feat(qboot)：重构算法注册和流处理机制 - 引入了 `qboot_algo.c` 来处理算法操作和注册事宜。 - 更新了 `qboot_fal_ops.c`，在读取、擦除和写入函数中使用 `rt_uint32_t` 来表示偏移量和长度。 - 创建了 `qboot_ops.c` 来管理头文件解析和 I/O 操作。 - 添加了 `qboot_stream.c` 来处理固件包的流传输和解压缩。
