+++ inc/qboot_algo.h
@@ -15,7 +15,7 @@
 #ifndef __QBOOT_ALGO_H__
 #define __QBOOT_ALGO_H__
 
-#include "qboot.h"
+#include "qboot_stream.h"
 
 #define QBOOT_ALGO_CRYPT_NONE  0
 #define QBOOT_ALGO_CRYPT_XOR   1
@@ -39,9 +39,11 @@
 
 typedef struct
 {
-    rt_err_t (*init)(void);                                 /**< Optional initializer for decryption. */
-    rt_err_t (*decrypt)(rt_uint8_t *out, const rt_uint8_t *in, size_t len); /**< Decrypt input buffer of @p len bytes into @p out. */
-    rt_err_t (*deinit)(void);                               /**< Optional cleanup called after decrypt stage. */
+    rt_err_t (*init)(void);                     /**< Optional initializer for decryption. */
+    rt_err_t (*decrypt)(rt_uint8_t *out,        /**< Output buffer for plaintext. */
+                        const rt_uint8_t *in,   /**< Input ciphertext buffer. */
+                        rt_uint32_t len);       /**< Number of bytes to decrypt. */
+    rt_err_t (*deinit)(void);                   /**< Optional cleanup called after decrypt stage. */
 } qboot_crypto_ops_t;
 
 /**
@@ -49,39 +51,28 @@ typedef struct
  */
 typedef struct
 {
-    rt_err_t (*init)(void);                                         /**< Optional initializer for decompression. */
-    rt_err_t (*decompress)(const qbt_stream_buf_t *buf,             /**< Input/output buffers. */
-                           qbt_stream_status_t *out,                /**< Stream results. */
-                           const qbt_stream_ctx_t *ctx);            /**< Stream context for this call. */
-    rt_err_t (*deinit)(void);                                       /**< Optional cleanup after decompression. */
+    rt_err_t (*init)(void);                                   /**< Optional initializer for decompression. */
+    rt_err_t (*decompress)(const qbt_stream_buf_t *buf,       /**< Input/output buffers. */
+                           qbt_stream_status_t *out,          /**< [out] Stream results. */
+                           const qbt_stream_ctx_t *ctx);      /**< Stream context for this call. */
+    rt_err_t (*deinit)(void);                                 /**< Optional cleanup after decompression. */
 } qboot_cmprs_ops_t;
 
-typedef struct
+typedef struct qboot_algo_ops
 {
-    rt_uint16_t algo_id;                                            /**< Algorithm identifier (compression/encryption). */
+    rt_uint16_t algo_id;                                    /**< Algorithm identifier (compression/encryption). */
     const qboot_crypto_ops_t *crypt;                        /**< NULL when no decryption stage required. */
     const qboot_cmprs_ops_t *cmprs;                         /**< NULL when no decompression stage required. */
-    rt_err_t (*apply)(void *src_handle, void *dst_handle,   /**< Optional apply hook for differential updates. */
-                      fw_info_t *fw_info, size_t patch_offset);
+    rt_err_t (*apply)(void *src_handle,                     /**< Source handle to read from. */
+                      void *dst_handle,                     /**< Destination handle to write to. */
+                      fw_info_t *fw_info,                   /**< Firmware header context. */
+                      rt_uint32_t patch_offset);            /**< Patch data offset within the package. */
 } qboot_algo_ops_t;
 
-/**
- * @brief Register algorithm ops for decrypt/compress handlers.
- *
- * The table size is auto-sized by the supported crypto/cmprs values. Each entry
- * is identified either by the compression algo id or the encryption algo id when
- * compression is not present. Each registered entry must provide a non-NULL
- * `cmprs->deprocess`. Passing the same algo twice returns -RT_ERROR.
- */
 rt_err_t qboot_algo_register(const qboot_algo_ops_t *ops, rt_uint16_t algo_id);
-
-/**
- * @brief Find algorithm table entry.
- *
- * @param crypt_id Encryption algorithm id (QBOOT_ALGO_CRYPT_*).
- * @param cmprs_id Compression algorithm id (QBOOT_ALGO_CMPRS_*).
- * @return ops pointer or NULL when not registered.
- */
-const qboot_algo_ops_t *qboot_algo_find(rt_uint16_t algo_id);
+const qboot_algo_ops_t *qbt_fw_get_algo_ops(const fw_info_t *fw_info, rt_uint16_t *out_algo_id);
+rt_bool_t qbt_fw_algo_init(const qboot_algo_ops_t *algo_ops);
+void qbt_fw_algo_deinit(const qboot_algo_ops_t *algo_ops);
+rt_err_t qbot_algo_startup(void);
 
 #endif
