+++ inc/qboot_cfg.h
@@ -34,6 +34,15 @@
 #define QBOOT_RELEASE_SIGN_ALIGN_SIZE 8//can is 4, 8, 16
 #define QBOOT_RELEASE_SIGN_WORD       0x5555AAAA
 
+#define QBOOT_BUF_SIZE 4096//must is 4096
+#if (defined(QBOOT_USING_QUICKLZ) || defined(QBOOT_USING_FASTLZ))
+#define QBOOT_CMPRS_READ_SIZE 4096 //it can is 512, 1024, 2048, 4096,
+#define QBOOT_CMPRS_BUF_SIZE  (QBOOT_BUF_SIZE + QBOOT_CMPRS_READ_SIZE + 32)
+#else
+#define QBOOT_CMPRS_READ_SIZE QBOOT_BUF_SIZE
+#define QBOOT_CMPRS_BUF_SIZE  QBOOT_BUF_SIZE
+#endif
+
 #ifdef QBOOT_USING_STATUS_LED
 #ifndef QBOOT_STATUS_LED_PIN
 #define QBOOT_STATUS_LED_PIN 0
